app/requirements.txt
flask==3.0.3
flask-limiter==3.8.0
prometheus-flask-exporter==0.23.1
gunicorn==22.0.0
python-dotenv==1.0.1